refactor: remove duplicate SHEL20001 game setup surface
Drop the unused CtGameSetup full-screen layout and consolidate new-game
setup on DLG10001 (NewGameLeaderSelectionDialog) across code, specs,
Widgetbook, tests, and l10n.
