+++ SPEC/game/game-setup.md
@@ -8,7 +8,7 @@ Pre-game phases that configure, generate, and populate the game world before tur
 
 **Phase order:** Config → World Generation → **Old World ownership (Great Powers + Minor Nations; locked assigner when the locked full-init profile and OW partition/role gates pass, otherwise BFS growth per pipeline)** → **New World tribe assignment (locked assigner on NW remainders; map+setup retries on documented topology failures)** → Faction & Initial State → Capital-Choice Phase.
 
-**Human/AI slot assignment:** Which Great Power slots are human vs AI is configurable at creation via `GameSetupConfig.humanGreatPowerSlotIndices` (`Set<int>`, default `{0}` = slot 0 / `gp1` human, the rest AI). An empty set creates a fully-AI game (used by the Full-AI observer). Slot indices are validated against `[0, greatPowerCount)`; out-of-range indices fail setup. The player-app new-game flow leaves the default (one human, slot 0) — see `SPEC/ui/game-setup.md`. Mechanics live in [game-setup-pipeline.md](../program/game-setup-pipeline.md) § Human/AI slot assignment.
+**Human/AI slot assignment:** Which Great Power slots are human vs AI is configurable at creation via `GameSetupConfig.humanGreatPowerSlotIndices` (`Set<int>`, default `{0}` = slot 0 / `gp1` human, the rest AI). An empty set creates a fully-AI game (used by the Full-AI observer). Slot indices are validated against `[0, greatPowerCount)`; out-of-range indices fail setup. The player-app new-game flow leaves the default (one human, slot 0) — see `SPEC/ui/new-game-leader-selection-dialog.md`. Mechanics live in [game-setup-pipeline.md](../program/game-setup-pipeline.md) § Human/AI slot assignment.
 
 **Tuned AI profiles (Refs #3444):** `GameSetupConfig.aiProfileByGpId` maps AI Great Power id → blessed profile name (`null` or absent entry = normal hardcoded personality). Chosen in the new-game leader dialog per AI slot; copied onto `Game.aiProfileByGpId` at game creation. Only personality parameter categories apply at runtime (not `victory_config`).
